+++ packages/atomic-elements/src/components/Content/List/List.context.ts
@@ -1,0 +1,4 @@
+import { createComponentContext } from "@utils/index";
+import { ListProps } from ".";
+
+export const ListContext = createComponentContext<ListProps>();