+++ packages/atomic-elements/src/components/Content/List/List.stories.tsx
@@ -1,0 +1,159 @@
+import { Meta, StoryObj } from "@storybook/react";
+import { List } from ".";
+import { MarginArgTypes, RenderPropsArgTypes } from "@sb/helpers";
+import { Text } from "../../Typography/Text";
+import { Flex } from "../../Layout/Flex";
+import { View } from "@components/Layout/View";
+import { IconButton } from "@components/Buttons/IconButton";
+import { CheckBox } from "@components/Inputs/Checkbox";
+import { getCssProps } from "@sb/cssprops";
+
+export default {
+  title: "Content/List",
+  component: List,
+  subcomponents: { "List.Item": List.Item },
+  parameters: {
+    cssprops: getCssProps("List"),
+  },
+  argTypes: {
+    ...RenderPropsArgTypes,
+    ...MarginArgTypes,
+  },
+} as Meta<typeof List>;
+
+type Story = StoryObj<typeof List>;
+
+export const Primary: Story = {
+  args: {
+    children: [
+      <List.Item key="4">
+        <Text as="div" $size="3" $weight="bold">
+          Medicine
+        </Text>
+        <Text $size="2" $color="neutral500">
+          Description for medicine. Lorem ipsum dolor sit amet, consectetur
+          adipiscing elit.
+        </Text>
+      </List.Item>,
+      <List.Item key="5">
+        <Text as="div" $size="3" $weight="bold">
+          Travelling
+        </Text>
+        <Text $size="2" $color="neutral500">
+          Description for travelling. Lorem ipsum dolor sit amet, consectetur
+          adipiscing elit.
+        </Text>
+      </List.Item>,
+      <List.Item key="6">
+        <Text as="div" $size="3" $weight="bold">
+          Gaming
+        </Text>
+        <Text $size="2" $color="neutral500">
+          Description for gaming. Lorem ipsum dolor sit amet, consectetur
+          adipiscing elit.
+        </Text>
+      </List.Item>,
+    ],
+  },
+};
+
+export const WithActions: Story = {
+  args: {
+    children: [
+      <List.Item key="4">
+        <Flex $gap="2" $align="center">
+          <View $flexGrow="1">
+            <Text as="div" $size="3" $weight="bold">
+              Medicine
+            </Text>
+            <Text $size="2" $color="neutral500">
+              Description for medicine. Lorem ipsum dolor sit amet, consectetur
+              adipiscing elit.
+            </Text>
+          </View>
+          <IconButton icon="edit" variant="ghost" />
+          <IconButton icon="close" variant="ghost" />
+        </Flex>
+      </List.Item>,
+      <List.Item key="5">
+        <Flex $gap="2" $align="center">
+          <View $flexGrow="1">
+            <Text as="div" $size="3" $weight="bold">
+              Travelling
+            </Text>
+            <Text $size="2" $color="neutral500">
+              Description for travelling. Lorem ipsum dolor sit amet,
+              consectetur adipiscing elit.
+            </Text>
+          </View>
+          <IconButton icon="edit" variant="ghost" />
+          <IconButton icon="close" variant="ghost" />
+        </Flex>
+      </List.Item>,
+      <List.Item key="6">
+        <Flex $gap="2" $align="center">
+          <View $flexGrow="1">
+            <Text as="div" $size="3" $weight="bold">
+              Gaming
+            </Text>
+            <Text $size="2" $color="neutral500">
+              Description for gaming. Lorem ipsum dolor sit amet, consectetur
+              adipiscing elit.
+            </Text>
+          </View>
+          <IconButton icon="edit" variant="ghost" />
+          <IconButton icon="close" variant="ghost" />
+        </Flex>
+      </List.Item>,
+    ],
+  },
+};
+
+export const WithSelection: Story = {
+  args: {
+    children: [
+      <List.Item key="4">
+        <Flex $gap="2" $align="center">
+          <CheckBox aria-label="Select medicine" defaultSelected />
+          <View $flexGrow="1">
+            <Text as="div" $size="3" $weight="bold">
+              Medicine
+            </Text>
+            <Text $size="2" $color="neutral500">
+              Description for medicine. Lorem ipsum dolor sit amet, consectetur
+              adipiscing elit.
+            </Text>
+          </View>
+        </Flex>
+      </List.Item>,
+      <List.Item key="5">
+        <Flex $gap="2" $align="center">
+          <CheckBox aria-label="Select travelling" />
+          <View $flexGrow="1">
+            <Text as="div" $size="3" $weight="bold">
+              Travelling
+            </Text>
+            <Text $size="2" $color="neutral500">
+              Description for travelling. Lorem ipsum dolor sit amet,
+              consectetur adipiscing elit.
+            </Text>
+          </View>
+        </Flex>
+      </List.Item>,
+      <List.Item key="6">
+        <Flex $gap="2" $align="center">
+          <CheckBox aria-label="Select gaming" defaultSelected />
+          <View $flexGrow="1">
+            <Text as="div" $size="3" $weight="bold">
+              Gaming
+            </Text>
+            <Text $size="2" $color="neutral500">
+              Description for gaming. Lorem ipsum dolor sit amet, consectetur
+              adipiscing elit.
+            </Text>
+          </View>
+        </Flex>
+      </List.Item>,
+    ],
+  },
+};